requirements.txt
# Python dependencies for slam_dense_reconstruction

# Core dependencies
open3d>=0.17.0
numpy>=1.21.0
opencv-python>=4.5.0
pyrealsense2>=2.50.0
pyyaml>=6.0

# Optional dependencies
matplotlib>=3.5.0  # For visualization
tqdm>=4.60.0       # Progress bars
pytest>=7.0.0      # Testing
